banking: types fixed

+++ Kernel/simple.c
@@ -22,7 +22,7 @@ int pagemap_alloc(ptptr p)
 
 int pagemap_realloc(uint16_t size)
 {
-  if (size + PROGBASE >= (uint16_t) ramtop)
+  if (size + (uint16_t)PROGBASE >= (uint16_t) ramtop)
     return ENOMEM;
   return 0;
 }
